requirements.txt
vk_api==11.9.9
python-telegram-bot==13.15
schedule==1.2.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
yt-dlp==2023.11.16
ffmpeg-python==0.2.0 